Added to AccountInfo and more
Changed minimum password length to 8, removed phone authentication
because it's not needed and we don't have a SMS server. Made Account
info visible when logged-in only, and added some info to the page, but
don't know how to get the real data in. Someone just has to add the data
bindings to grab the data.